Composite Raw Decoder: ref: CompositeRawDecoder & jsonlinedecoder
