+++ unit_tests/sources/declarative/decoders/test_json_decoder.py
@@ -8,14 +8,9 @@ import os
 import pytest
 import requests
 
-from airbyte_cdk import YamlDeclarativeSource
-from airbyte_cdk.models import SyncMode
 from airbyte_cdk.sources.declarative.decoders import GzipJsonDecoder
 from airbyte_cdk.sources.declarative.decoders.json_decoder import JsonDecoder, JsonlDecoder
-from airbyte_cdk.sources.declarative.models import DeclarativeStream as DeclarativeStreamModel
-from airbyte_cdk.sources.declarative.parsers.model_to_component_factory import (
-    ModelToComponentFactory,
-)
+
 
 
 @pytest.mark.parametrize(
@@ -62,67 +57,6 @@ def large_event_response_fixture():
             file.write(jsonl_string)
     yield (lines_in_response, file_path)
     os.remove(file_path)
-
-
-@pytest.mark.slow
-@pytest.mark.limit_memory("20 MB")
-def test_jsonl_decoder_memory_usage(requests_mock, large_events_response):
-    lines_in_response, file_path = large_events_response
-    content = """
-    name: users
-    type: DeclarativeStream
-    retriever:
-      type: SimpleRetriever
-      decoder:
-        type: JsonlDecoder
-      paginator:
-        type: "NoPagination"
-      requester:
-        path: "users/{{ stream_slice.slice }}"
-        type: HttpRequester
-        url_base: "https://for-all-mankind.nasa.com/api/v1"
-        http_method: GET
-        authenticator:
-          type: NoAuth
-        request_headers: {}
-        request_body_json: {}
-      record_selector:
-        type: RecordSelector
-        extractor:
-          type: DpathExtractor
-          field_path: []
-      partition_router:
-        type: ListPartitionRouter
-        cursor_field: "slice"
-        values:
-          - users1
-          - users2
-          - users3
-          - users4
-    primary_key: []
-        """
-
-    factory = ModelToComponentFactory()
-    stream_manifest = YamlDeclarativeSource._parse(content)
-    stream = factory.create_component(
-        model_type=DeclarativeStreamModel, component_definition=stream_manifest, config={}
-    )
-
-    def get_body():
-        return open(file_path, "rb", buffering=30)
-
-    counter = 0
-    requests_mock.get("https://for-all-mankind.nasa.com/api/v1/users/users1", body=get_body())
-    requests_mock.get("https://for-all-mankind.nasa.com/api/v1/users/users2", body=get_body())
-    requests_mock.get("https://for-all-mankind.nasa.com/api/v1/users/users3", body=get_body())
-    requests_mock.get("https://for-all-mankind.nasa.com/api/v1/users/users4", body=get_body())
-
-    stream_slices = list(stream.stream_slices(sync_mode=SyncMode.full_refresh))
-    for stream_slice in stream_slices:
-        for _ in stream.retriever.read_records(records_schema={}, stream_slice=stream_slice):
-            counter += 1
-
-    assert counter == lines_in_response * len(stream_slices)
 
 
 @pytest.mark.parametrize(
